src/FinanceBot/requirements.txt
python-telegram-bot==13.13
filelock==3.12.2